flip to have `index.d.ts` be the complete types definition, and import from that file into `src`, to prevent imports in the definitions file
